Move greedy files around
