+++ bfit/greedy/__init__.py
@@ -19,4 +19,3 @@
 # along with this program; if not, see <http://www.gnu.org/licenses/>
 #
 # ---
-from bfit.greedy.greedy_strat import GreedyStrategy
