+++ bfit/test/test_greedy_utils.py
@@ -25,9 +25,9 @@ r"""Test file for 'bfit.greedy.greedy_utils'. """
 import numpy as np
 import numpy.testing as npt
 
-from bfit.greedy.greedy_utils import check_redundancies, get_next_choices
-from bfit.greedy.greedy_utils import get_next_possible_coeffs_and_exps2
-from bfit.greedy.greedy_utils import get_two_next_choices, pick_two_lose_one
+from bfit.greedy_utils import check_redundancies, get_next_choices
+from bfit.greedy_utils import get_next_possible_coeffs_and_exps2
+from bfit.greedy_utils import get_two_next_choices, pick_two_lose_one
 
 
 __all__ = ["test_check_redundancies", "test_get_next_possible_coeffs_and_exps",
